+++ modules/__init__.py
@@ -2,10 +2,7 @@
 from .GPSCheck import GPS_Check
 from .PhotoTable import Stats
 from .DateTime import DateTime
-from .Flash import Flash_Chart
-from .Focal import Focal_Chart
-from .PieChart import Pi_Chart
+from .PieChart import PieChart
 
 __all__ = ["GPS_Check", "Stats", "DateTime",
-           "Flash_Chart", "Focal_Chart",
-           "Pi_Chart"]
+           "PieChart"]